+++ src/main/java/gov/usgs/wma/waterdata/collections/CollectionsDao.java
@@ -10,16 +10,9 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public class CollectionsDao extends SqlSessionDaoSupport {
-	public static final String NULL_TIME_SERIES = "\"timeSeries\":null";
-	public static final String EMPTY_TIME_SERIES= "\"timeSeries\":[]";
-
-	private CollectionParams collectionsParams;
-	
-	
 	@Autowired
-	public CollectionsDao(SqlSessionFactory sqlSessionFactory, CollectionParams collectionsParams) {
+	public CollectionsDao(SqlSessionFactory sqlSessionFactory) {
 		setSqlSessionFactory(sqlSessionFactory);
-		this.collectionsParams = collectionsParams;
 	}
 
 	public String getCollectionsJson(Map<String, Object> params) {
@@ -48,17 +41,6 @@ public class CollectionsDao extends SqlSessionDaoSupport {
 
 	public String getDiscreteDataObsverationsJson(Map<String, Object> params) {
 		return getSqlSession().selectOne("observations.getDiscreteDataObsverationsJson", params);
-	}
-
-	public String getStatisticalTimeSeries(String collectionId, String featureId) {
-		Map<String,Object> params = collectionsParams.buildParams(collectionId, featureId);
-		
-		String json = getSqlSession().selectOne("collections.getStatisticalTimeSeriesJson", params);
-		
-		if (json != null && json.contains(NULL_TIME_SERIES)) {
-			json = json.replace(NULL_TIME_SERIES, EMPTY_TIME_SERIES);
-		}
-		return json;
 	}
 
 }