+++ src/test/java/gov/usgs/wma/waterdata/collections/CollectionsControllerIT.java
@@ -30,6 +30,7 @@ import com.github.springtestdbunit.annotation.DatabaseSetup;
 
 @SpringBootTest(webEnvironment=WebEnvironment.RANDOM_PORT)
 @DatabaseSetup("classpath:/testData/monitoringLocation/")
+@DatabaseSetup("classpath:/testData/groundwaterDailyValue/")
 public class CollectionsControllerIT extends BaseCollectionsIT {
 	@Autowired
 	private TestRestTemplate restTemplate;
@@ -169,6 +170,12 @@ public class CollectionsControllerIT extends BaseCollectionsIT {
 
 	@Test
 	public void collectionNotFoundItemsTest() {
+		String rtn = doCollectionRequest("/collections/xyz/items", HttpStatus.NOT_FOUND);
+		assertEquals(ogc404Payload, rtn);
+	}
+
+	@Test
+	public void collectionNotFoundFeatureExistsTest() {
 		String rtn = doCollectionRequest("/collections/xyz/items/USGS-07227448", HttpStatus.NOT_FOUND);
 		assertEquals(ogc404Payload, rtn);
 	}