+++ src/test/java/gov/usgs/wma/waterdata/springinit/FileSensingDataSetLoader.java
@@ -2,7 +2,13 @@ package gov.usgs.wma.waterdata.springinit;
 
 import java.io.InputStream;
 import java.sql.Date;
+import java.text.SimpleDateFormat;
 import java.time.Instant;
+import java.time.LocalDate;
+import java.time.format.DateTimeFormatter;
+import java.util.AbstractMap;
+import java.util.ArrayList;
+import java.util.List;
 
 import org.dbunit.dataset.IDataSet;
 import org.dbunit.dataset.ReplacementDataSet;
@@ -11,6 +17,8 @@ import org.dbunit.dataset.xml.FlatXmlDataSetBuilder;
 import org.springframework.core.io.Resource;
 
 import com.github.springtestdbunit.dataset.AbstractDataSetLoader;
+
+import static gov.usgs.wma.waterdata.springinit.BaseIT.getReplacementTokens;
 
 public class FileSensingDataSetLoader extends AbstractDataSetLoader {
 
@@ -37,7 +45,15 @@ public class FileSensingDataSetLoader extends AbstractDataSetLoader {
 
 	private IDataSet createReplacementDataSet(IDataSet iDataSet) {
 		ReplacementDataSet replacementDataSet = new ReplacementDataSet(iDataSet);
-		replacementDataSet.addReplacementObject("[today]", Date.from(Instant.now()));
+
+		List<AbstractMap.SimpleEntry<String, String>> replacements = BaseIT.getReplacementTokens();
+		for (AbstractMap.SimpleEntry<String, String> replacement: replacements) {
+			String orgText = replacement.getKey();
+			String newText = replacement.getValue();
+
+			replacementDataSet.addReplacementSubstring(orgText, newText);
+		}
+
 		return replacementDataSet;
 	}
 }